Moved ES queries into their own file
