+++ flask/app.py
@@ -3,93 +3,8 @@ from flask_cors import CORS,cross_origin
 from elasticsearch import Elasticsearch, helpers
 import uuid
 from game import Game
+from queries import *
 import os
-
-#Searches title field for given phrase
-def search_title_match_phrase(es, query, size=10):
-    resp = es.search( index="articles",
-    body={
-        "query": {
-            "match_phrase": {
-                "title": {
-                    "query": query}
-            }
-        }
-    },
-    size=size
-    )
-    return resp
-
-def search_title_match(es, query):
-    resp = es.search( index="articles",
-    body={
-        "query": {
-            "match": {
-                "title": {
-                    "query": query}
-            }
-        }
-    },
-    )
-    return resp
-
-def search_title_exact(es, query):
-    resp = es.search( index="articles",
-    body={
-        "query": {
-            "term": {
-                "title.keyword": query
-            }
-        }
-    },
-    )
-    return resp
-
-#Get 10 random articles
-def search_random(es):
-    resp = es.search( index="articles",
-    body={
-        "query": {
-            "function_score": {
-                "random_score": {}
-            }
-        }
-    },
-    )
-    return resp
-
-def search_similar(es, query):
-    resp = es.search(index = "articles",
-    body={
-        "query": {
-            "more_like_this": {
-                "fields": ["text"],
-                "like": query,
-                "min_term_freq": 1,
-                "min_doc_freq": 1    
-            }
-        }
-    },
-    )
-    return resp
-
-#Get the article with _id = id
-def search_id(es, id):
-    return es.get(index="articles", id=id)
-
-#Get the article with ATTRIBUTE id = id
-def search_id_match(es, id):
-    resp = es.search( index="articles",
-    body={
-        "query": {
-            "match": {
-                "id": {
-                    "query": id}
-            }
-        }
-    },
-    )
-    return resp
 
 app = Flask(__name__)
 cors = CORS(app, resources={r"/*": {"origins": "*"}})
@@ -102,43 +17,16 @@ es = Elasticsearch(
 #Create a dictionary of game_id: Game
 games = dict()
 
-#An example response containing the Python Wikipedia page
-article_python = {
-    "title": "Python (programming language)",
-    "source": "Wikipedia",
-    "text": r"""Python is a high-level, general-purpose programming language. Its design philosophy emphasizes code readability with the use of significant indentation.[31]
-Python is dynamically typed and garbage-collected. It supports multiple programming paradigms, including structured (particularly procedural), object-oriented and functional programming. It is often described as a "batteries included" language due to its comprehensive standard library.[32][33]
-Guido van Rossum began working on Python in the late 1980s as a successor to the ABC programming language and first released it in 1991 as Python 0.9.0.[34] Python 2.0 was released in 2000. Python 3.0, released in 2008, was a major revision not completely backward-compatible with earlier versions. Python 2.7.18, released in 2020, was the last release of Python 2.[35]
-Python consistently ranks as one of the most popular programming languages, and has gained widespread use in the machine learning community.[36][37][38][39]"""
-}
-
-#An example response containing the C Wikipedia page
-article_c = {
-    "title": "C (programming language)",
-    "source": "Wikipedia",
-    "text": r"""C (pronounced /ˈsiː/ – like the letter c)[6] is a general-purpose computer programming language. It was created in the 1970s by Dennis Ritchie, and remains very widely used and influential. By design, C's features cleanly reflect the capabilities of the targeted CPUs. It has found lasting use in operating systems, device drivers, and protocol stacks, but its use in application software has been decreasing.[7] C is commonly used on computer architectures that range from the largest supercomputers to the smallest microcontrollers and embedded systems.
-A successor to the programming language B, C was originally developed at Bell Labs by Ritchie between 1972 and 1973 to construct utilities running on Unix. It was applied to re-implementing the kernel of the Unix operating system.[8] During the 1980s, C gradually gained popularity. It has become one of the most widely used programming languages,[9][10] with C compilers available for practically all modern computer architectures and operating systems. The book The C Programming Language, co-authored by the original language designer, served for many years as the de facto standard for the language.[11][1] C has been standardized since 1989 by the American National Standards Institute (ANSI) and the International Organization for Standardization (ISO).
-"""
-}
+#Define endpoints for Flask API
 
 #This function runs when a GET request is sent to 127.0.0.1:{port}
 @app.route("/")
 def index() -> str:
     #Render some html with a link when a user visits the site with no path
     if len(games) > 0:
-        return f"There are {len(games)} games currently running. Navigate to <a href=\"/serve_article/{list(games.keys())[0]}\"> /serve_article/{list(games.keys())[0]} </a> to see an example response. <br />Navigate to <a href=\"/search_article/{list(games.keys())[0]}\"> /search_article/{list(games.keys())[0]} </a> to search the database. <br />Navigate to <a href=\"/start_article/{list(games.keys())[0]}\"> /start_article/{list(games.keys())[0]} </a> to get a random start article.<br /> <form action = \"/api/start_game\" method =\"post\"> <input type=\"submit\" value = \"Create New Game\"></form>" 
+        return f"There are {len(games)} games currently running. <br />Navigate to <a href=\"/search_article/{list(games.keys())[0]}\"> /search_article/{list(games.keys())[0]} </a> to search the database. <br />Navigate to <a href=\"/start_article/{list(games.keys())[0]}\"> /start_article/{list(games.keys())[0]} </a> to get a random start article.<br /> <form action = \"/api/start_game\" method =\"post\"> <input type=\"submit\" value = \"Create New Game\"></form>" 
     else:
         return "There are no games currently running. <br /> <form action = \"/api/start_game\" method =\"post\"> <input type=\"submit\" value = \"Create New Game\"></form>"     
-
-#This function runs when a GET request is sent to 127.0.0.1:{port}/serve_article/{id}
-@app.get("/serve_article/<uuid:id>")
-def serve_article_get(id: uuid.UUID) -> dict:
-    #Check if the given id matches the player
-    if id in games:
-        #Serve an example article
-        return article_python
-    else:
-        return jsonify({"error": "The provided id does not match a valid game id."})
 
 #This function runs when a POST request is sent to 127.0.0.1:{port}/serve_article/{id}
 @app.post("/serve_article/<uuid:id>")
